+++ packages/flutter_mimir/lib/src/ffi/io.dart
@@ -1,6 +1,10 @@
 import 'dart:ffi';
 import 'dart:io';
 
+import 'package:meta/meta.dart';
+
+@internal
+// ignore: public_member_api_docs
 DynamicLibrary createLibraryImpl() {
   const base = 'embedded_milli';